fix error in parser/parser_isbs.c and correct libft

+++ libft/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jekim <arabi1549@naver.com>                +#+  +:+       +#+        */
+/*   By: jekim <jekim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/01 01:02:23 by jekim             #+#    #+#             */
-/*   Updated: 2021/10/10 14:50:59 by jekim            ###   ########.fr       */
+/*   Updated: 2021/12/11 00:28:35 by jekim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ char	*ft_substr(char const *src, unsigned int start, size_t len)
 	ret = (char *)malloc(sizeof(char) * (len + 1));
 	if (!ret)
 		return (NULL);
-	while (ix < len)
+	while (ix < (int)len)
 		ret[ix++] = src[start++];
 	ret[ix] = '\0';
 	return (ret);
